backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
pytz
sqlparse
psycopg2-binary
python-dotenv